requirements.txt
pyyaml==5.3.1
selenium==3.141.0
